chore: refactor address to break up large file
