+++ sites/public/pages/applications/contact/address.tsx
@@ -15,13 +15,10 @@ import {
   FieldGroup,
   ProgressNav,
   t,
-  MultiLineAddress,
-  Address,
 } from "@bloom-housing/ui-components"
 import FormsLayout from "../../../layouts/forms"
 import { useState } from "react"
 import { useForm } from "react-hook-form"
-import GeocodeService from "@mapbox/mapbox-sdk/services/geocoding"
 import { Select } from "@bloom-housing/ui-components/src/forms/Select"
 import { PhoneField } from "@bloom-housing/ui-components/src/forms/PhoneField"
 import {
@@ -33,56 +30,11 @@ import {
 } from "@bloom-housing/shared-helpers"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
-
-interface FoundAddress {
-  newAddress?: Address
-  originalAddress?: Address
-  invalid?: boolean
-}
-
-const findValidatedAddress = (
-  address: Address,
-  setFoundAddress: React.Dispatch<React.SetStateAction<FoundAddress>>,
-  setNewAddressSelected: React.Dispatch<React.SetStateAction<boolean>>
-) => {
-  const geocodingClient = GeocodeService({
-    accessToken: process.env.mapBoxToken || process.env.MAPBOX_TOKEN,
-  })
-
-  geocodingClient
-    .forwardGeocode({
-      query: `${address.street}, ${address.city}, ${address.state}, ${address.zipCode}`,
-      limit: 1,
-    })
-    .send()
-    .then((response) => {
-      const [street, city, region] = response.body.features[0].place_name.split(", ")
-      const regionElements = region.split(" ")
-      const zipCode = regionElements[regionElements.length - 1]
-
-      if (!zipCode) {
-        setNewAddressSelected(false)
-        setFoundAddress({ invalid: true, originalAddress: address })
-      } else {
-        setNewAddressSelected(true)
-        setFoundAddress({
-          originalAddress: address,
-          newAddress: {
-            street,
-            street2: address.street2 && address.street2 !== "" ? address.street2 : undefined,
-            city,
-            state: address.state,
-            zipCode,
-          },
-        })
-      }
-    })
-    .catch((err) => {
-      console.error(`Error calling Mapbox API: ${err}`)
-      setNewAddressSelected(false)
-      setFoundAddress({ invalid: true, originalAddress: address })
-    })
-}
+import {
+  FoundAddress,
+  findValidatedAddress,
+  AddressValidationSelection,
+} from "../../../src/forms/applications/ValidateAddress"
 
 const ApplicationAddress = () => {
   const [verifyAddress, setVerifyAddress] = useState(false)
@@ -614,82 +566,9 @@ const ApplicationAddress = () => {
           </div>
 
           {verifyAddress && (
-            <>
-              <div className="form-card__group">
-                {foundAddress.newAddress && (
-                  <fieldset>
-                    <legend className="field-note mb-4">
-                      {t("application.contact.suggestedAddress")}
-                    </legend>
-
-                    <div className="field field--inline">
-                      <input
-                        type="radio"
-                        name="chooseaddress"
-                        id="foundaddress"
-                        value="found"
-                        checked={newAddressSelected}
-                        onChange={(e) => setNewAddressSelected(e.target.checked)}
-                      />
-                      <label htmlFor="foundaddress" className="font-alt-sans font-semibold">
-                        <MultiLineAddress
-                          address={{
-                            street: foundAddress.newAddress.street2
-                              ? `${foundAddress.newAddress.street}<br/>${foundAddress.newAddress.street2}`
-                              : foundAddress.newAddress.street,
-                            city: foundAddress.newAddress.city,
-                            state: foundAddress.newAddress.state,
-                            zipCode: foundAddress.newAddress.zipCode,
-                          }}
-                        />
-                      </label>
-                    </div>
-                  </fieldset>
-                )}
-                {foundAddress.invalid && <p>[couldn't find a verified address]</p>}
-                {foundAddress.originalAddress && (
-                  <fieldset className="mt-6">
-                    <legend className="field-note mb-4">
-                      {t("application.contact.youEntered")}
-                    </legend>
-
-                    <div className="flex items-start">
-                      <div className="field field--inline">
-                        <input
-                          type="radio"
-                          name="chooseaddress"
-                          id="originaladdress"
-                          value="original"
-                          checked={!newAddressSelected}
-                          onChange={(e) => setNewAddressSelected(!e.target.checked)}
-                        />
-                        <label htmlFor="originaladdress" className="font-alt-sans font-semibold">
-                          <MultiLineAddress
-                            address={{
-                              street: foundAddress.originalAddress.street2
-                                ? `${foundAddress.originalAddress.street}<br/>${foundAddress.originalAddress.street2}`
-                                : foundAddress.originalAddress.street,
-                              city: foundAddress.originalAddress.city,
-                              state: foundAddress.originalAddress.state,
-                              zipCode: foundAddress.originalAddress.zipCode,
-                            }}
-                          />
-                        </label>
-                      </div>
-                      <Button
-                        unstyled
-                        className="font-alt-sans uppercase font-semibold mt-0 mr-0"
-                        onClick={() => {
-                          setVerifyAddress(false)
-                        }}
-                      >
-                        {t("t.edit")}
-                      </Button>
-                    </div>
-                  </fieldset>
-                )}
-              </div>
-            </>
+            <AddressValidationSelection
+              {...{ foundAddress, newAddressSelected, setNewAddressSelected, setVerifyAddress }}
+            />
           )}
 
           <div className="form-card__pager">
